Continued adventures in containerization
As this is now meant to run in a container, log error messages to the
console instead of a file.

This also fixes an issue with ssl certs: go would refuse to connect to
the Discord websocket because of an unknown cert. We now download the
certs in an Alpine container and add them to the final scratch one.